requirements.txt
alembic==0.9.2
amqp==2.1.4
astroid==1.5.2
Babel==2.4.0
bcrypt==3.1.3
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
dnspython==1.15.0
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.5
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
lazy-object-proxy==1.3.1
ldap3==2.2.4
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
pyasn1==0.2.3
pycparser==2.17
pylint==1.7.1
PyMySQL==0.7.11
python-dateutil==2.6.0
python-editor==1.0.3
python-etcd==0.4.5
pytz==2017.2
redis==2.10.5
requests==2.17.3
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-datatables==1.2.0
urllib3==1.21.1
vine==1.1.3
webassets==0.12.1
Werkzeug==0.12.2
wrapt==1.10.10
WTForms==2.1
xlrd==1.0.0